move file.ts to core/base from generator/base
ISSUE=#227
